Refactor: SearchBox 하위 영역 Area로 구분
