+++ FE/src/components/SearchBox/CalendarArea/CalendarCourousel/Calendar/style.ts
