+++ FE/src/components/SearchBox/CalendarArea/CalendarCourousel/constants.ts
