+++ FE/src/components/SearchBox/CalendarArea/CalendarCourousel/style.ts
