+++ FE/src/components/SearchBox/PeopleArea/index.tsx
@@ -1,0 +1,11 @@
+import * as S from '../style';
+
+export default function PeopleArea() {
+  return (
+    <>
+      <S.PeoplePart>
+        <S.Label>인원</S.Label>
+      </S.PeoplePart>
+    </>
+  );
+}
