+++ FE/src/components/SearchBox/PriceArea/PriceChart/Chart/index.tsx
@@ -5,7 +5,7 @@ import {
   calculateYRatio,
   fillContext,
   setPositions,
-} from '@/components/SearchBox/PriceChart/Chart/util';
+} from '@/components/SearchBox/PriceArea/PriceChart/Chart/util';
 import { PRIMARY_COLOR, SECONDARY_COLOR } from '@/style';
 
 export interface ChartProps {
