+++ FE/src/components/SearchBox/PriceArea/PriceChart/Chart/util.ts
