+++ FE/src/components/SearchBox/PriceArea/PriceChart/MultiRangerSlider/index.tsx
