+++ FE/src/components/SearchBox/PriceArea/PriceChart/MultiRangerSlider/style.ts
