+++ FE/src/components/SearchBox/PriceArea/PriceChart/PriceInfo/index.tsx
