+++ FE/src/components/SearchBox/PriceArea/PriceChart/constants.js
