+++ FE/src/components/SearchBox/PriceArea/PriceChart/index.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 
-import Chart from '@/components/SearchBox/PriceChart/Chart';
-import { CANVAS_HEIGHT, CANVAS_WIDTH } from '@/components/SearchBox/PriceChart/constants';
-import MultiRangerSlider from '@/components/SearchBox/PriceChart/MultiRangerSlider';
-import PriceInfo from '@/components/SearchBox/PriceChart/PriceInfo';
+import Chart from '@/components/SearchBox/PriceArea/PriceChart/Chart';
+import { CANVAS_HEIGHT, CANVAS_WIDTH } from '@/components/SearchBox/PriceArea/PriceChart/constants';
+import MultiRangerSlider from '@/components/SearchBox/PriceArea/PriceChart/MultiRangerSlider';
+import PriceInfo from '@/components/SearchBox/PriceArea/PriceChart/PriceInfo';
 import { PriceChartDataProps } from '@/data/chartData';
 
 import * as S from './style';
