+++ FE/src/components/SearchBox/PriceArea/PriceChart/style.ts
